Add files via upload
Adding more Java source files, as well as JSON files that contain format-specific resources that are used by the Java code.

These "CreateSpecialContent...." files are unique to the output format requested by the user, when the
AUTHOR program is executed. Each Java object is dynamically loaded, and it in turn, loads the JSON file that it
uses.

The JSON contains output formatting fragments used by the Java code. Using this system works well for
making changes to the output format, without having to recompile the Java code. Such fragments include
those for HTML portions and tags, for instance.

Note (sadly) that some output is too complex to use fragments kept on a JSON file. If you like what you get
in the output, great. If you want to "tweak" the formatting, you must look at the JSON file for the
desired format. In addition (again, sadly) you may have to actually look at the Java source code.