Add files via upload
Initial add of 2 "sinks", these for the FOP output. The programs create the ".fo" files
used by Apache FOP. Apache FOP then creates the PDF files of the document.

AUTHOR uses a "sink/source" pattern. "Source" is the handler for the AUTHOR text input,
which for now is "BookCreate.java". Each "sink" follows a specific interface, and
it loaded dynamically at runtime to make the document output.

There is a "sink" for each output format type. Because of this pattern, it
is quite possible to make new Sink's for any desired document format.