Rename reveal.html to en_route_example.html
Change name. This is the output of AUTHOR for this presentation
example. Your output should be quite similar. This example
can be viewed in your browser, as long as the various
suppport directories. Those include the "pics", "thumbs", "css",
and "js" directories. Also, this presentation requires the presence
of the "utility.js" file in the current directory.