Add files via upload
Add Perl script that creates the repeated content for the 1,000 page example. Invoking
the script with the "perl" command will create a standard output stream that
contains the repeated information.

The "x1000.auth" file contains, as the first few lines, the needed
boilerplate and introduction for the example.

You can alter the loop parameters on the Perl script to make a larger
or small example AUTHOR file.